# src/server/api/go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Luminox.",
        "title": "Luminox API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/disk": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all disks under a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disk"
                ],
                "summary": "List disks",
                "parameters": [
                    {
                        "type": "string",
                        "example": "alice@luminox.io",
                        "description": "User identifier to filter disks",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit of disks to return, default 20. Max 200.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor for pagination. Use the cursor from the previous response to get the next page.",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "Order by created_at descending if true, ascending if false (default false)",
                        "name": "time_desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.ListDisksOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# List disks\ndisks = client.disks.list(limit=10, time_desc=True)\nfor disk in disks.items:\n    print(f\"Disk: {disk.id}\")\n\n# List disks for a specific user\ndisks = client.disks.list(user='alice@luminox.io', limit=10)\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// List disks\nconst disks = await client.disks.list({ limit: 10, timeDesc: true });\nfor (const disk of disks.items) {\n  console.log(`Disk: ${disk.id}`);\n}\n\n// List disks for a specific user\nconst userDisks = await client.disks.list({ user: 'alice@luminox.io', limit: 10 });\n"
                    }
                ]
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a disk group under a project. Optionally associate with a user identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disk"
                ],
                "summary": "Create disk",
                "parameters": [
                    {
                        "description": "CreateDisk payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateDiskReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Disk"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Create a disk\ndisk = client.disks.create()\nprint(f\"Created disk: {disk.id}\")\n\n# Create a disk for a specific user\ndisk = client.disks.create(user='alice@luminox.io')\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Create a disk\nconst disk = await client.disks.create();\nconsole.log(`Created disk: ${disk.id}`);\n\n// Create a disk for a specific user\nconst userDisk = await client.disks.create({ user: 'alice@luminox.io' });\n"
                    }
                ]
            }
        },
        "/disk/{disk_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a disk by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "disk"
                ],
                "summary": "Delete disk",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Disk ID",
                        "name": "disk_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                },
                "x-code-samples": [
                    {
                        "label": "Python",
                        "lang": "python",
                        "source": "from luminox import LuminoxClient\n\nclient = LuminoxClient(api_key='sk_project_token')\n\n# Delete a disk\nclient.disks.delete(disk_id='disk-uuid')\n"
                    },
                    {
                        "label": "JavaScript",
                        "lang": "javascript",
                        "source": "import { LuminoxClient } from '@luminox/luminox';\n\nconst client = new LuminoxClient({ apiKey: 'sk_project_token' });\n\n// Delete a disk\nawait client.disks.delete('disk-uuid');\n"
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "handler.CreateDiskReq": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "string",
                    "example": "alice@luminox.io"
                }
            }
        },
...
        "model.Disk": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "service.ListDisksOutput": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Disk"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Project Bearer token (e.g., \"Bearer sk-ac-xxxx\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}